refactor: change the file names to temporary file names
