+++ client/src/pages/temp_login/temp_login.tsx
@@ -3,7 +3,7 @@ import dattLogo from  '../../assets/images/logo.png'
 import {FaUser, FaEye} from 'react-icons/fa';
 import { useDispatch } from 'react-redux'
 import { login } from '../../reducers/user/userSlice'
-import { fetchApi } from '../../components/Localhost/Localhost'
+import { fetchApi } from '../../components/Localhost/temp_localhost'
 import { jwtDecode } from "jwt-decode";
 
 interface LoginFormProps {
